+++ huginn-net-http/src/lib.rs
@@ -186,16 +186,3 @@ impl<'a> HuginnNetHttp<'a> {
         }
     }
 }
-
-/// Represents the output from HTTP analysis.
-///
-/// This struct contains various optional outputs that can be derived
-/// from analyzing HTTP packets.
-#[derive(Debug)]
-pub struct HttpAnalysisResult {
-    /// Information derived from HTTP request packets.
-    pub http_request: Option<HttpRequestOutput>,
-
-    /// Information derived from HTTP response packets.
-    pub http_response: Option<HttpResponseOutput>,
-}
