+++ huginn-net-tcp/src/lib.rs
@@ -180,22 +180,3 @@ impl<'a> HuginnNetTcp<'a> {
         }
     }
 }
-
-/// Represents the output from TCP analysis.
-///
-/// This struct contains various optional outputs that can be derived
-/// from analyzing TCP packets.
-#[derive(Debug)]
-pub struct TcpAnalysisResult {
-    /// Information derived from SYN packets.
-    pub syn: Option<SynTCPOutput>,
-
-    /// Information derived from SYN-ACK packets.
-    pub syn_ack: Option<SynAckTCPOutput>,
-
-    /// Information about the Maximum Transmission Unit (MTU).
-    pub mtu: Option<MTUOutput>,
-
-    /// Information about the system uptime.
-    pub uptime: Option<UptimeOutput>,
-}
